+++ src/components/help/help.js
@@ -87,7 +87,7 @@ const Help = TooltipDecorator(class Help extends React.Component{
     return (
       <a className={ this.mainClasses } href={ this.props.href }>
         <Icon
-          type='info'
+          type='help'
           tooltipMessage={ this.props.children }
           tooltipPosition={ this.props.tooltipPosition }
           tooltipAlign={ this.props.tooltipAlign }