+++ src/utils/decorators/input-label/input-label.js
@@ -78,7 +78,7 @@ let InputLabel = (ComposedComponent) => class Component extends ComposedComponen
     return classNames(
       'common-input__label', {
         'common-input__label--inline': this.props.labelInline,
-        'common-input__label--help': this.props.fieldHelp,
+        'common-input__label--help': this.props.labelHelp,
         'common-input__label--align-right': this.props.labelAlign === 'right'
       }
     );