+++ CHANGELOG.md
@@ -4,7 +4,17 @@
 
 * Profile - User to show portrait with name and email.
 * AppWrapper - confines your content to the width of your application.
+* Menu
 * NavigationBar
+
+## Input Label Padding
+
+* All input label padding has been slightly increased.
+
+## Help Updates
+
+* Help component has been updated with a new icon.
+* Input Label decorator has been fixed to render the help class for labelHelp.
 
 ## Acronymize Function